Rename http integration tests to transport
