+++ truss/_integration-tests/Makefile
@@ -3,24 +3,24 @@ OK_COLOR=\e[38;5;118m
 UNDER=\n________________________________________________________________________________\n
 END_COLOR_LINE=$(UNDER)$(NO_COLOR)
 
-HTTPTEST_MSG=\n$(OK_COLOR)Starting http end to end test:$(END_COLOR_LINE)
+TRANSPORT_TEST_MSG=\n$(OK_COLOR)Starting transport end to end test:$(END_COLOR_LINE)
 
-CLITEST_MSG=\n$(OK_COLOR)Start server and cliclient generate, build, and run test:$(END_COLOR_LINE)
+CLI_TEST_MSG=\n$(OK_COLOR)Start server and cliclient generate, build, and run test:$(END_COLOR_LINE)
 
 
 all: test
 
-test: clean test-http test-cli
+test: clean test-transport test-cli
 
-test-http:
-	@printf '$(HTTPTEST_MSG)'
-	$(MAKE) -C http
+test-transport:
+	@printf '$(TRANSPORT_TEST_MSG)'
+	$(MAKE) -C transport
 
 test-cli:
-	@printf '$(CLITEST_MSG)'
+	@printf '$(CLI_TEST_MSG)'
 	go test -v ./cli
 
 clean:
 	go test ./cli -clean
-	$(MAKE) -C http clean
+	$(MAKE) -C transport clean
