+++ truss/_integration-tests/transport/Makefile
@@ -5,7 +5,7 @@ END_COLOR_LINE=$(UNDER)$(NO_COLOR)
 
 TRUSS_MSG=\n$(OK_COLOR)Running Truss...$(END_COLOR_LINE)
 
-TEST_RUNNING_MSG=\n$(OK_COLOR)Running http tests:$(END_COLOR_LINE)
+TEST_RUNNING_MSG=\n$(OK_COLOR)Running transport tests:$(END_COLOR_LINE)
 
 TRUSS_AGAIN_MSG=\n$(OK_COLOR)Running Truss... again, to test regeneration$(END_COLOR_LINE)
 
@@ -13,16 +13,15 @@ all: test
 
 test:
 	@echo -e '$(TRUSS_MSG)'
-	truss httptest.proto
-	cp -r handlers httptest-service
+	truss transport-test.proto
+	cp -r handlers transport-service
 	@echo -e '$(TEST_RUNNING_MSG)'
 	go test -v
 	@echo -e '$(TRUSS_AGAIN_MSG)'
-	truss httptest.proto
+	truss transport-test.proto
 	@echo -e '$(TEST_RUNNING_MSG)'
 	go test -v
 	$(MAKE) clean
 
 clean:
-	rm -rf httptest-service
-	rm -rf third_party
+	rm -rf transport-service
